backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
motor==3.3.2
pymongo==4.6.0
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
opencv-python==4.8.1.78
scikit-image==0.21.0
tensorflow==2.14.0
torch==2.1.0
torchvision==0.16.0
scikit-learn==1.3.2
numpy==1.24.3
scipy==1.11.3
pillow==10.1.0
matplotlib==3.8.1
celery==5.3.4
pydantic==2.5.0
pydantic-settings==2.1.0
aiofiles==23.2.1
python-dotenv==1.0.0
structlog==23.2.0
prometheus-client==0.19.0
slowapi==0.1.9
bcrypt==4.0.1